modified: refactored db dirs
